Apply first refactoring round
