+++ artifacts/compile_commands.json.tmpl
@@ -31,6 +31,11 @@
 },
 {
   "directory": "{{PROJECT_PATH}}/build",
+  "command": "{{CONDA_PREFIX}}/bin/clang++ -DGFLAGS_IS_A_DLL=0 -DGLOG_CUSTOM_PREFIX_SUPPORT -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS -I{{PROJECT_PATH}}/src -fvisibility-inlines-hidden -std=c++20 -fmessage-length=0 -march=nocona -mtune=haswell -ftree-vectorize -fPIC -fstack-protector-strong -fno-plt -O2 -ffunction-sections -pipe -isystem {{CONDA_PREFIX}}/include -O3 -DNDEBUG -fprofile-instr-generate -fcoverage-mapping -std=c++20 -o CMakeFiles/arx.dir/src/codegen/ast-to-stdout.cpp.o -c {{PROJECT_PATH}}/src/codegen/ast-to-stdout.cpp",
+  "file": "{{PROJECT_PATH}}/src/codegen/ast-to-stdout.cpp"
+},
+{
+  "directory": "{{PROJECT_PATH}}/build",
   "command": "{{CONDA_PREFIX}}/bin/clang++ -DGFLAGS_IS_A_DLL=0 -DGLOG_CUSTOM_PREFIX_SUPPORT -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS -I{{PROJECT_PATH}}/src -fvisibility-inlines-hidden -std=c++20 -fmessage-length=0 -march=nocona -mtune=haswell -ftree-vectorize -fPIC -fstack-protector-strong -fno-plt -O2 -ffunction-sections -pipe -isystem {{CONDA_PREFIX}}/include -O3 -DNDEBUG -fprofile-instr-generate -fcoverage-mapping -std=c++20 -o CMakeFiles/arx.dir/src/utils.cpp.o -c {{PROJECT_PATH}}/src/utils.cpp",
   "file": "{{PROJECT_PATH}}/src/utils.cpp"
 },
@@ -93,6 +98,11 @@
   "directory": "{{PROJECT_PATH}}/build",
   "command": "{{CONDA_PREFIX}}/bin/clang++ -DGFLAGS_IS_A_DLL=0 -DGLOG_CUSTOM_PREFIX_SUPPORT -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS -I{{PROJECT_PATH}}/src -I{{PROJECT_PATH}}/build/tests -fvisibility-inlines-hidden -std=c++20 -fmessage-length=0 -march=nocona -mtune=haswell -ftree-vectorize -fPIC -fstack-protector-strong -fno-plt -O2 -ffunction-sections -pipe -isystem {{CONDA_PREFIX}}/include -fsanitize=address -O1 -fno-omit-frame-pointer -fprofile-instr-generate -fcoverage-mapping -O3 -DNDEBUG -fprofile-instr-generate -fcoverage-mapping -std=c++20 -o tests/CMakeFiles/ArxTests.dir/__/src/codegen/ast-to-object.cpp.o -c {{PROJECT_PATH}}/src/codegen/ast-to-object.cpp",
   "file": "{{PROJECT_PATH}}/src/codegen/ast-to-object.cpp"
+},
+{
+  "directory": "{{PROJECT_PATH}}/build",
+  "command": "{{CONDA_PREFIX}}/bin/clang++ -DGFLAGS_IS_A_DLL=0 -DGLOG_CUSTOM_PREFIX_SUPPORT -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS -I{{PROJECT_PATH}}/src -I{{PROJECT_PATH}}/build/tests -fvisibility-inlines-hidden -std=c++20 -fmessage-length=0 -march=nocona -mtune=haswell -ftree-vectorize -fPIC -fstack-protector-strong -fno-plt -O2 -ffunction-sections -pipe -isystem {{CONDA_PREFIX}}/include -fsanitize=address -O1 -fno-omit-frame-pointer -fprofile-instr-generate -fcoverage-mapping -O3 -DNDEBUG -fprofile-instr-generate -fcoverage-mapping -std=c++20 -o tests/CMakeFiles/ArxTests.dir/__/src/codegen/ast-to-stdout.cpp.o -c {{PROJECT_PATH}}/src/codegen/ast-to-stdout.cpp",
+  "file": "{{PROJECT_PATH}}/src/codegen/ast-to-stdout.cpp"
 },
 {
   "directory": "{{PROJECT_PATH}}/build",
