+++ meson.build
@@ -2,10 +2,32 @@ project('arx', 'cpp', 'c',
   license : 'Apache-2.0',
   version : '1.5.0',  # semantic-release
   default_options : [
-    'warning_level=3',
+    'warning_level=everything',
     'cpp_std=c++20',
   ]
 )
+
+add_global_arguments(
+  '-Wpedantic',
+  '-Wno-c++98-compat-pedantic',
+  '-Wno-padded',
+  '-Wno-missing-prototypes',
+  '-Wshadow',
+  '-Wnon-virtual-dtor',
+  '-Wcast-align',
+  '-Wunused',
+  '-Woverloaded-virtual',
+  '-Wconversion',
+  '-Wsign-conversion',
+  '-Wdouble-promotion',
+  '-Wformat=2',
+  '-Wimplicit-fallthrough',
+  '-Wsuggest-override',
+  '-Wnull-dereference',
+  '-Wold-style-cast',
+  language : 'cpp'
+)
+
 
 PROJECT_PATH = meson.source_root()
 
@@ -25,7 +47,7 @@ project_src_files = files(
   # note: note working yet
   # PROJECT_PATH + '/src/codegen/ast-to-llvm.cpp',
   PROJECT_PATH + '/src/codegen/ast-to-object.cpp',
-  PROJECT_PATH + '/src/codegen/ast-to-output.cpp',
+  PROJECT_PATH + '/src/codegen/ast-to-stdout.cpp',
   PROJECT_PATH + '/src/error.cpp',
   PROJECT_PATH + '/src/io.cpp',
   PROJECT_PATH + '/src/lexer.cpp',
